// models/models.tsx
export interface IMetaDetails {
  title: string;
  description: string;
  keyword: string;
  ogTitle: string;
}
// ...
export interface IProductDetails {
  id: number;
  price: string;
  displayImage: IImage;
  offerPrice: string;
  ecomProductId: string;
  ecomModelCode: string;
  type: string;
  category?: string;
  slug: string;
  title: string;
  subTitle: string;
  locale: string;
  localizations: ILocalization[];
  megaMenuThumb: IImage;
  banner: ITopBanner[];
  specs: IKeyValue[];
  design: IDesign[];
  detailItems: DetailItem[];
  gallery: IImage[];
  colorVariants: IColorVariant[];
  interiorColorVariants: IColorVariant[];
  trims: ITrims[];
  convenience: IConvenience[];
  buildLink: string;
  sliderImage: IImage;
  showInMainMenu: boolean;
  homePageSliderSpec: IHomePageSliderSpec;
  metaDetails?: IMetaDetails;
}

export interface IHomePageSliderSpec {
  spec1: string;
  spec2: string;
}
// ...
export interface IConvenience {
  title: string;
  description: string;
  image: IImage;
}

export interface ITrims {
  title: string;
  subTitle: string;
  price: string;
  infoText: string;
  spec: ISpecText[];
  image: IImage;
}

export interface ISpecText {
  spefcItem: string;
}

export interface IColorVariant {
  title: string;
  colorIcon: IImage;
  image: IImage;
}
export interface DetailItem {
  title: string;
  subtitle?: any;
  items: Item[];
}

export interface Item {
  id: number;
  title: string;
  description: string;
  image: IImage;
}

export interface IHighlights {
  id: number;
  title: string;
  subtitle: string;
  items: Item[];
}

export interface IKeyValue {
  label: string;
  value: string;
}

export interface IImage {
  id: number;
  name: string;
  url: string;
}

export interface IDesign {
  id: number;
  title: string;
  items: IImageWithDescription[];
}

export interface IImageWithDescription {
  id: number;
  title: string;
  description: string;
  image: IImage;
}

export interface ILocalization {
  locale: string;
  id: string;
  slug: string;
}
// ...
export interface IMotorcycleLanding {
  banner: ITopBanner[];
  bottomBanner1: ITopBanner;
  bottomBanner2: ITopBanner;
  bottomBanner3: ITopBanner;
  metaDetails: IMetaDetails;
}
// ...
export interface ITopBanner {
  title: string;
  description: string;
  type: string;
  image: {
    url: string;
    formats: IimageFormat;
  };
  automobile: IRelationData;
  motorcycle: IRelationData;
  marine: IRelationData;
  power_product: IRelationData;
  button: IButton[];
  relativeUrl: string;
  absoluteUrl: string;
}

export interface IimageFormat {
  small: { url: string };
}

export interface IRelationData {
  title: string;
  slug: string;
  locale: string;
  localizations: ILocalization[];
}
export interface IButton {
  color: "light_color" | "theme_color";
  text: string;
  relativeUrl: string;
  absoluteUrl: string;
  automobile: IRelationData;
  motorcycle: IRelationData;
  marine: IRelationData;
  power_product: IRelationData;
}
export interface IWelcomeText {
  title: string;
  description: string;
}
export interface IShoppingTool {
  title: string;
  description: string;
  items: IShoppingToolItem[];
}
export interface IShoppingToolItem extends ITopBanner {
  link: string;
}
export interface ISocialMedia {
  title: string;
  items: ISocialMediaItem[];
}
export interface ISocialMediaItem {
  url: string;
  iconClass: string;
  position: string;
  image: {
    url: string;
  };
}
// ...
export interface ILinkItem {
  label: string;
  relativeUrl?: string;
  absoluteUrl?: string;
}

export interface IFooterMenu {
  title: string;
  items: ILinkItem[];
  isDisable?: boolean;
}
export interface IFooter {
  title: string;
  copyRight: string;
  menuItems: IFooterMenu[];
  socialMedia: ISocialMedia[];
  serviceLinks: IFooterMenu[];
  connectWithUs: IFooterMenu[];
  extraLinks: IFooterMenu[];
}